Extract Position
